Fix user_defined and allow multiple reactants/products (#277)
This PR makes the following two changes to how the user-defined `@rxn`
reactions are handled within the code:

- fixes wrong variable in `user_defined()`

There was a line that incorrectly refers to `product` instead of
`reactant` when handling user-defined degradation reactions (e.g., `@rxn
A -> 0:...`), which led to the placeholder variable (`"0"`) being
included as one of the species in the resulting model.

- Adds feature to allow multiple reactants and/or products in
user-defined reactions

Currently, the `@rxn` reaction type only accepts single reactant and
product reactions. Changes were made to allow users to describe custom
reactions with multiple reactants/products by concatenating them with a
`+` symbol within the Text2Model file.

The implementation of this feature mimics that of the
`state_transition()` function.

I will attach a toy Text2Model file that I used for testing below.

I would appreciate any type of comments/suggestions. Thanks in advance!

```
@rxn 0 -> A: p[kf_a] | kf_a=0.1
@rxn A -> 0: p[deg_a] * u[A] | deg_a=0.05
@rxn A + B -> AB + B: p[kf_ab] * u[A] * u[B] | kf_ab=0.1
@rxn AB + B -> 0: p[deg_ab] * u[AB] | deg_ab=0.05

@obs A: u[A]
@obs B: u[B]
@obs AB: u[AB]

@sim tspan:[0, 100]
@sim condition control: init[A]=1; init[B]=1
```